Refactoring and improving house generation
Changed the way that the renderer handles room children, so that all of the logic for how a room should be rendered happens for each room, instead of a lot of stuff happening when processing all the children. Hopefully this will be easier to read and debug.

Also, I'm now cycling all possible directions for every room instead of just trying one random one.. Because of the way that I handle houses being shifted north and east, this tends to mean that there is more empty space created since the tiles are shifted even if the room can't be places (if another room is in the way for instance), but I like the effect, it makes it feel like the house has a yard.